+++ middleware-arena-account/account.dto/src/main/java/com/mware/account/dto/response/AccountBalanceResponse.java
@@ -6,8 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 /**
  * 账户余额响应对象（对外暴露，仅含展示所需字段，不含 id / 审计时间）。
  */
@@ -19,5 +17,6 @@ public class AccountBalanceResponse {
 
     private Long userId;
 
-    private BigDecimal balance;
+    /** 账户余额（单位：分） */
+    private Long balance;
 }
